+++ cmd/crucible/main.go
@@ -115,6 +115,7 @@ func newRootCmd(stdout, stderr io.Writer) *cobra.Command {
 		newServiceCmd(opts),
 		newLogsCmd(opts),
 		newImageCmd(opts),
+		newAppCmd(opts),
 		newBuildCmd(opts),
 		newSnapshotCmd(opts),
 		newForkCmd(opts),